Introduced new AbstractUsernamePasswordAuthenticationFilter
