+++ web/src/main/java/org/springframework/security/web/authentication/UsernamePasswordAuthenticationFilter.java
@@ -44,9 +44,10 @@ import org.springframework.util.Assert;
  * @author Ben Alex
  * @author Colin Sampaleanu
  * @author Luke Taylor
+ * @author Marcel Overdijk
  * @since 3.0
  */
-public class UsernamePasswordAuthenticationFilter extends AbstractAuthenticationProcessingFilter {
+public class UsernamePasswordAuthenticationFilter extends AbstractUsernamePasswordAuthenticationFilter {
     //~ Static fields/initializers =====================================================================================
 
     public static final String SPRING_SECURITY_FORM_USERNAME_KEY = "username";
@@ -54,21 +55,11 @@ public class UsernamePasswordAuthenticationFilter extends AbstractAuthentication
 
     private String usernameParameter = SPRING_SECURITY_FORM_USERNAME_KEY;
     private String passwordParameter = SPRING_SECURITY_FORM_PASSWORD_KEY;
-    private boolean postOnly = true;
-
-    //~ Constructors ===================================================================================================
-
-    public UsernamePasswordAuthenticationFilter() {
-        super(new AntPathRequestMatcher("/login","POST"));
-    }
 
     //~ Methods ========================================================================================================
 
-    public Authentication attemptAuthentication(HttpServletRequest request, HttpServletResponse response) throws AuthenticationException {
-        if (postOnly && !request.getMethod().equals("POST")) {
-            throw new AuthenticationServiceException("Authentication method not supported: " + request.getMethod());
-        }
-
+    @Override
+    protected UsernamePasswordAuthenticationToken obtainUsernamePasswordAuthenticationToken(HttpServletRequest request) {
         String username = obtainUsername(request);
         String password = obtainPassword(request);
 
@@ -82,12 +73,7 @@ public class UsernamePasswordAuthenticationFilter extends AbstractAuthentication
 
         username = username.trim();
 
-        UsernamePasswordAuthenticationToken authRequest = new UsernamePasswordAuthenticationToken(username, password);
-
-        // Allow subclasses to set the "details" property
-        setDetails(request, authRequest);
-
-        return this.getAuthenticationManager().authenticate(authRequest);
+        return new UsernamePasswordAuthenticationToken(username, password);
     }
 
     /**
@@ -119,17 +105,6 @@ public class UsernamePasswordAuthenticationFilter extends AbstractAuthentication
     }
 
     /**
-     * Provided so that subclasses may configure what is put into the authentication request's details
-     * property.
-     *
-     * @param request that an authentication request is being created for
-     * @param authRequest the authentication request object that should have its details set
-     */
-    protected void setDetails(HttpServletRequest request, UsernamePasswordAuthenticationToken authRequest) {
-        authRequest.setDetails(authenticationDetailsSource.buildDetails(request));
-    }
-
-    /**
      * Sets the parameter name which will be used to obtain the username from the login request.
      *
      * @param usernameParameter the parameter name. Defaults to "username".
@@ -147,18 +122,6 @@ public class UsernamePasswordAuthenticationFilter extends AbstractAuthentication
     public void setPasswordParameter(String passwordParameter) {
         Assert.hasText(passwordParameter, "Password parameter must not be empty or null");
         this.passwordParameter = passwordParameter;
-    }
-
-    /**
-     * Defines whether only HTTP POST requests will be allowed by this filter.
-     * If set to true, and an authentication request is received which is not a POST request, an exception will
-     * be raised immediately and authentication will not be attempted. The <tt>unsuccessfulAuthentication()</tt> method
-     * will be called as if handling a failed authentication.
-     * <p>
-     * Defaults to <tt>true</tt> but may be overridden by subclasses.
-     */
-    public void setPostOnly(boolean postOnly) {
-        this.postOnly = postOnly;
     }
 
     public final String getUsernameParameter() {
